web/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.30.0
fal-client==0.3.0
mangum==0.17.0
python-json-logger==2.0.7